% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arma_svd}
\alias{arma_svd}
\title{Singular value decomposition via RcppArmadillo}
\usage{
arma_svd(X, method = "standard")
}
\arguments{
\item{X}{matrix}

\item{method}{one of \code{"standard"} or \code{"dc"}}
}
\description{
Singular value decomposition via RcppArmadillo
}

