% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cginv.R
\name{cginv}
\alias{cginv}
\alias{gginv}
\title{Generalized Inverse of a Matrix using a custom SVD}
\usage{
cginv(X, tol = sqrt(.Machine$double.eps), svd = NULL)

gginv(svd = NULL)
}
\arguments{
\item{X}{
Matrix for which the Moore-Penrose inverse is required.
}

\item{tol}{
A relative tolerance to detect zero singular values.
}

\item{svd}{A function that computes the singular value decomposition of a
  matrix}
}
\value{
A MP generalized inverse matrix for X.
}
\description{
The \code{cginv} function
calculates the Moore-Penrose generalized inverse of a matrix X using a custom
implementation for computing the singular value decomposition.
The \code{gginv} function creates a function with formals identical
to those of \code{MASS::\link[MASS]{ginv}} that uses a given SVD implementation.
}
\details{
The \code{svd} parameter is expected to adhere to the interface of
\code{base::\link[base]{svd}}. It will be called as \code{svd(x)} (with the
\code{nu} and \code{nv} arguments unset) and is expected to return a named
list with components \code{d}, \code{u} and \code{v}.
}
\author{
Authors of the \code{MASS} package
}
\seealso{
\code{\link[MASS]{ginv}}, \code{\link[base]{svd}}
}

