% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grake-package.r
\docType{package}
\name{grake-package}
\alias{grake}
\alias{grake-package}
\title{grake: Generalized Raking as Described by Deville et al. (1993)}
\description{
This package implements functions for generalized raking as
described in Deville et al. (1993): "Generalized raking procedures in survey
sampling".  The code strives to provide performance and ease of use.
}
\seealso{
dss, gginv
}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org} [contributor]

Authors:
\itemize{
  \item Andreas Alfons
  \item Matthias Templ
}

Other contributors:
\itemize{
  \item Brian Ripley (gginv() has been derived from MASS::ginv()) [contributor]
}

}
